Rename the generic form input types to kebab-case
Nine of the input type files used snake_case while the rest of the folder and the repo use kebab-case. Renames and import updates only, no change in behaviour.
