+++ scripts/core/ui/components/generic-form/form-field.tsx
@@ -2,21 +2,21 @@ import React from 'react';
 import {PlainText} from './input-types/plain-text';
 import {assertNever} from 'core/helpers/typescript-helpers';
 import {isIFormGroup, isIFormField, GenericFormFieldType} from './interfaces/form';
-import {VocabularySingleValue} from './input-types/vocabulary_single_value';
+import {VocabularySingleValue} from './input-types/vocabulary-single-value';
 import {TextEditor3} from './input-types/text-editor3';
 import {noop} from 'lodash';
 import {CheckboxInput} from './input-types/checkbox';
 import {ContentFilterSingleValue} from './input-types/content-filter-single-value';
 import {IInputType} from './interfaces/input-types';
-import {DeskSingleValue} from './input-types/desk_single_value';
-import {StageSingleValue} from './input-types/stage_single_value';
-import {getMacroSingleValue} from './input-types/macro_single_value';
+import {DeskSingleValue} from './input-types/desk-single-value';
+import {StageSingleValue} from './input-types/stage-single-value';
+import {getMacroSingleValue} from './input-types/macro-single-value';
 import {YesNo} from './input-types/yes-no';
 import {IFormField, IFormGroup} from 'superdesk-api';
-import {SelectMultipleValues} from './input-types/select_multiple_values';
+import {SelectMultipleValues} from './input-types/select-multiple-values';
 import {NumberComponent} from './input-types/number';
 import {Select} from './input-types/select';
-import {SelectAsync, SelectMultipleAsync} from './input-types/select_async';
+import {SelectAsync, SelectMultipleAsync} from './input-types/select-async';
 import {DurationComponent} from './input-types/duration';
 import {AlertInput} from './input-types/alert';
 import {ReadonlyCopyableText} from './input-types/readonly-copiable-text';
