+++ scripts/core/ui/components/generic-form/input-types/select-single-value.tsx
