+++ scripts/core/ui/components/generic-form/input-types/select.tsx
@@ -1,4 +1,4 @@
-import {getSelectSingleValue} from './select_single_value';
+import {getSelectSingleValue} from './select-single-value';
 
 export const Select = getSelectSingleValue(
     (props) => Promise.resolve(props.formField.component_parameters.options),
